fix basket issue and local storage - final one

+++ webapp/src/pages/Checkout/index.tsx
@@ -4,21 +4,28 @@ import useProduct from '../../hooks/useProduct';
 
 const Checkout = () => {
   const { productInfo } = useProduct();
-  if (!productInfo) return <></>;
+  console.log(productInfo);
+
+  const truthyDataParsed = productInfo && Object.values(productInfo || '').every((item) => item);
+
+  if (!productInfo || !truthyDataParsed) return <></>;
+
   return (
     <div className="h-screen">
       <img src={productInfo.product?.image} />
       {productInfo.price}
       {productInfo.product?.title}
-      <div>
-        <Button
-          text="Clear shopping Basket"
-          onClick={() => {
-            localStorage.clear();
-            window.location.reload();
-          }}
-        />
-      </div>
+      {truthyDataParsed && (
+        <div>
+          <Button
+            text="Clear shopping Basket"
+            onClick={() => {
+              localStorage.clear();
+              window.location.reload();
+            }}
+          />
+        </div>
+      )}
     </div>
   );
 };
